Complicate UserSearch to prefer kb users rather than sbs for dupes
